Create VOLO
This song was written after trying and succeeding to drink underage at the bar of the namesake several times, as I had gotten in with the bands that were playing, and on one fateful night having vivid projection of ejection from the premises. I am unsure if this tale is true or false, many of the sentiments expressed are nonfictional however there is some historical fiction involved as the teenage soul has fondness of. I was 16 going on 17 years old, which was the beginning of my intense unmanageable periods of active addiction as is elucidated in the song, and truth be told I was probably heavily under the influence when writing this or pensively recovering from a bender. I wrote this song by placing my hands on the middle c of the piano and going down one three finger chord on the keys at a time, four times to create the rythm of the song. Flourishes were added ad lib as I learned how the piano works more formally, though it was entirely intuitive and spontaneous in writitg. The original sheet of line paper it was written on to this day has no chords or note markings written on it, as I did not know all the chords I was playing until today when I synthesized the composition for this documentation.