Create I AM A GOD
This song was written shortly before my first trip to rehab when I was 16 I remember because I played it for all the girls there in the adolescent rehab and they went nuts over it. I remember being able to make myself cry playing it, from the spirit most times and preformatively other times. I always hated the werewolf verse but other people have liked it so I never altered the song or edited it much from its original version. As is characteristic of most of my early songwriting there is no bridge, and the chord progression is identical throughout the verses and chorous. My first opus. 

NOTE: Between now and the next hard copy of a song there is a large gap, there was productivity from this time but nothing that has traveled with me and stood the test of time. I still have some audio recordings of these long lost songs but they scarcely exist on paper except for fragments of poetry shuffled away in some folder at my childhood home.