Create EXCUSES (etc.)
This was written in the peak of my disputing not all men feminist phase of my life. The song is a compilation of excuses I have heard in my life sentence by sentence. It was written to be catchy and simple. One of the shortest least verbose of the songs that I have written. It was written in two sections, in 2015 the point up to "she said she was 18" and from then on written in 2018.