Create EAGLE 20's
This song was written in my first voluntary treatment stay at BRC and was chronologically the first one I wrote in there. It was the day Jonathan R got kicked out of the rehab for taking other peoples Suboxone. I had created a bond with him that would later turn into a world renowned druggieship, I never knew if I liked him romantically or platonically or was so repulsed that I was attracted. I know we spent nights manically babbling about our experiences with heroin and dreams of traveling, well at least they were my dreams and he emphatically consented to be a part of them in the future. We planned a beautiful train ride to California but when I showed up in Philadelphia after treatment to meet him and get high we never made it off the alphabet streets in Kensington, unless it was to go to Camden to cop over the tracks. I loved him. He loved me enough to help kill me on one occasion after a subsequent overdose on  some hot Philly fentanyl. This all came later and I suppose the fatal attraction I felt to him that I expressed with such longing in this song was a precursive intuition to the trauma we would later share. God bless you Jonathan, here's to all the invincible junkies in this world who've got no veins left, steal their fathers pocket watches, and don't even get high anymore just get well.