Create AFTER FX
This song is open for interpretation to who its about, I wrote it with my mother in mind for the last two verses, myself for the first two verses and Jonathan for the chorus. it is about the gnawing pain felt from the internal existential division of self that is experienced in active addiction.