Create SAMSARA
This song was written in an improvisational jam between Nicolas (Nicky) on guitar and me on vocals as well as him singing ad lib with me recording in written form on the back of a piece of art I had made in there. The original copy of this song is still on the back of that piece of art to this day. Nicky and I formally wrote two songs together and had many more jams, our flow was immaculate and our harmonization was made in heaven. This song is deeply interpretive and reflective, pensive and abstract even, as our union often was. I love him to this day with my whole heart and have followed his journey from the outside looking in ever since I got into recovery and he stayed out of it for the most part. He is currently incarcerated in some capacity as he often was, having just come out of prison and into the inpatient facility when fate brought us together. I will get his art tattooed on me one day and have a copy of it with a letter from him in my song book at home. Nicky if you're reading this je t'aime plus que toute le monde.