BRC
Written ad lib again by Nicky and I at BRC inpatient, he contributed mostly to the lyrics of the first section of the song and I finished up the latter half of the song at a later point in our stay. I have busked this song from PA to CA and its a fan favorite for sure.