Philly Summer
This is the only song that survived out of the period of time after BRC and subsequently through my first Philadelphia run with Jon boy. I wrote this I feel like after coming down somewhere, maybe California??? I had a mandolin at my disposal so this is definitely the correct era for that. The contents of this song portray the acknowledgement of the insanity of my disease at the time of my Philly escapades and the travel I did during that period.