Create GAME
This piece was written as response to an assignment by a therapist inpatient treatment at Hazelden for adolescents the rehab where I wanted to use the most in my youth. I was there for a total of four or five months the longest I had ever been in treatment up until the willows later in life. This is also the treatment center where I taught myself how to play guitar, as the next song in this sequence will bear relevance to. This song is about the manipulative and playing with life or death things that I did in my active addiction. The idea that life was just a game that I got to play and choose how to manipulate others and the truth and myself in the process.