Create I FEEL LIKE IM IN KINDERGARTEN
This song was written at 90 days clean after the Philly episode number one in Minnesota at baby Hazelden. It is the first song I wrote on guitar and I was so proud to have made progress enough to know four whole chords. For a long time after that I only played songs with the five or six chords that I knew. I preformed this at the inpatient talent show and played it for all the girls on the unit. I took on a mothering and experienced recovery persona of a person there and this song was an expression of how deep down I felt like a child who literally didn't know how to walk. Music has a way of bringing the truth to light. Brynne P helped me out during a creative block on verse four and five, the words in the song are hers almost verbatim.