Create maybe this time
This song was also written as a therapy assignment inpatient at baby HZD, a crowd favorite on the unit. Played as intended it has a jazzy swing to it and very dramatic ballad lyrics. You can observe the point in this song when the realization of the starkness of my reality hit, the fact that I did hit bottom in my addiction, was homeless, was praying to die with every shot and was in fact after all of this incapable of getting better for myself, anything for you. I don't know who the you was in this scenario, probably my mother or family who had sent me there.