Create Sir End Her
Written in treatment at Hazelden for babies for an assignment with the therapist on trauma and intimacy and surrender, hence the name sir end her. So clever right? 19 Year old me thought so too. I was playing around with scoping out less conventional chord combinations and my command of the piano was the only instrument I could experiment with that on. I had only just played babies first chords on guitar and the piano was always available. I spent hours if not days of my six month stay there at the piano in the basement, sometimes people would congregate and sometimes I was stoically alone at the keys. It was a very productive era that stay.