Create Je, Moi, Mon (Perspective)
One of my first self describing songs, I am the she in the song and identified as a she at the time. On the nose love song to myself pretty much, hence the title I Me Mine in French. Written in adult Hazelden in st Paul Minnesota where I was inspired constantly and playing music with Sam all the time on the piano. Also progressed in guitar here at this treatment facility. This song was written during the time before I was kicked out and moved to Chicago.