Create Parliamentary
This song was written by inspiration following a reading of William Burroughs and that is the theme of the song as well. Possibly my favorite chorus I've ever written, knowing silence is unsettling that's why babies sing before they speak. One of my favorite to sing and yet a bit challenging due to the unconventional minor chord composition. Meant to be played on piano and in my experience does not translate well to guitar.