Create Dad's Not Punk (Thats okay)
A song of disenchantment with the punk agenda and feeling like a washed up impostor. I wrote this song after meeting some real punks in adult Hazelden in st Paul and realizing that I had traded in my punk identity for a hippie free loving tie dye wearing cop out. The struggle of identity is a theme in many of my songs and this one most directly, outlining the reasons I qualified for the scene and yet washed up out of it at a certain point. One of the only true folk punk songs I ever wrote.