Create Queen Baby
Also written inpatient at Hazelden for adults, song title inspired by the AA pamphlet King Baby. Fuck you dr bob women can be baby royalty too. The song is about relating with other people, the second verse being about how me raised as a woman in relation to men always felt like men forced implication of sexual or romantic nature and my dream at the time was that I could convey and have my true meaning in relating with them understood. The refrain is one of my favorite choruses I've ever written to a song ever, slant rhymes, cryptic literate wording, proud of that. As well as conveying the sentiment of never feeling like my truest meaning was ever delivered in a way that was matched or landed in the ways I meant it to. This is also one of very few songs of mine that has a bridge, which was particularly difficult fo rme to write compositionally. When actively played, omitting some of the accent chords such as Dm is totally acceptable and I end up doing that half of the time anyway. The sentiment behind that section of the song is a self reflection on why my style of relating is flawed... especially considering the fact that I didn't know an ounce of who I really was at that time in my life, I felt like an impostor in everything I did, especially in the area of relationships.