Create R.S.V.P.
RSVP stands for Ryan Scott Van Pouke a dear friend of mine and mutual crush from treatment at the Hazelden outpatient center that died of an overdose in supportive living in St Paul Minnesota. I was in a treatment center in Chicago at the time that his funeral was, and he was from Chicago so his funeral was there. It just so lined up that I got kicked out of treatment on the same day as his funeral so I packed my bags and headed for the funeral home. I got there and was shooting dope the whole time in the bathroom, where right outside the door a mother was mourning her son thet had just died from the poison I was using. On the flight home after the funeral I overdosed on the plane and ended up in the hospital for some days in NC. When I got home I wrote this song about Ryan as I was still grieving him. I wonder if you wondered that before you died or if your reaper looked black and empty just like mine, was a connection between our two deaths, one more permanent than the other. Bless his soul