Create Soulbound
I wrote this song sitting on the porch swing at my grandmothers house in Pennsylvania less than ten days after my overdose, being around the family for Christmas and evaluating my entire existence. Its one of my few happy songs or at least uplifting and was written to include the harshness of reality and the relative fecundity of it all. Written in a very country folk style, its meant to evoke feelings of old wisdom that is always latent.