Create Break Up Son pt.  IIV
Song about quitting cigarettes for the millionth time, hence the title breakup song part IIV which I thought meant seventeen for some reason. But yeah this song is breakup song part 17 and is about my experience quitting smoking ciggarettes after overdosing on the plane in NC and promptly picking up packs of smokes on the drive home, where I only had 40% lung capacity and function because when else was I going to be able to buy smokes for five dollars a pack??? This one has become a crowd favorite for sure, it woulds like a song about a person without the context of it directly being about cigarettes.