Create Nerves: No Reason Why Now is Not the Time
this poem was written sitting in the common room at corning community college gazing out at the buzzing and trifling of all the 18 year old kids which seemed so far away at the ripe age of 21 when I wrote this. I was getting extra credits for my bachelors degree taking upper Level gender studies and philosophy courses and very inspired by that content at the time. A queer longing waxing philosophical piece of poetry. Not originally intended to be a song but loosely very loosely became one.