Create Mountains Crumble
Yet another song about the disillusionment with the human condition, many subsequent songs would be written about this material. and this chord progression beginning with the F7 was to be used in many future songs as well, like new chords are used. Definitely a pessimistic piece, especially with the chorus removed in the final version of the song. The chorus was:
escape your truth when mountains crumble, you can move them kid dr. Seuss told you so.
And it ws in CDGC so much more upbeat but that was removed in the final product.