Create Apocalypso
Written in the midst of the covid pandemic this song outlines a hope for revolutionary action spurred by the state of emergency that was so impotently handled by the US government. My hopes were after that there would be an impetus for universal socialized healthcare for all. Americans, universal basic income, or at the very least comprehensive rights for essential workers, but alas we reflect on this situation three years later as a epic vegetative failing of government and society itself.