Create Falafels For Brains: Deep Fried
This song was written fresh out of rehab at home for the first time in awhile, I was in a pretty profoundly depressed state but this song has hope as well in a melancholy way. If this is the end of all I knew, then we are finally enough. Peace in the ending basically is the meaning of this song. I alter playing depending on audiences, and leave out the part about rehab and stories and skip to the last three verses instead. I am and always have been dissatisfied with the ending of this song, but alas this is how ive always played it.