Create Line Go Down
This is marked on the paper as a split LP, MAISIE X BOLO. It was written as part of the covid series when I was in college and covid first hit and we had to go totally online. Maisie was one of my favorite people that I bonded with and met her in a gender studies class. We wrote this song together apart over the internet and texting and sending voice messages to each other, she came up with the licks and chords and sent them to me and I did the words that I have documented. The nature of this is pretty direct, a folk punk song about capitalist pigs and hoping the market crash with covid would actually hurt them as much as it hurt the people on the ground who were actually suffering.