Create Shadoworkingonit
Song written in the midst of the only abusive relationship ive ever been in with Sam. He wanted my love and my light and my healing to his shadows and insanity and yet rejected with disgust and shame every expression of my shadow side. He was deeply disgusted with the fact that I am an addict, and yet contiuned to feed me drugs that he felt were acceptable for me. I played this for him and he said nothing in reflection essentially. This relationship ended when I went back to heroin and talked to him while I was high and he spewed vitriol at me for doing so, which neither of what we did was acceptable. The verbal abuse continued for months, and I ultimately ended up overdosing and dying on my childhood bedroom floor where my mother found me. This was my last time using, which was November of this same year. This relationship made It clear to me of why I never want to be a drug dealer, a narcissist, or be in a relationship with someone who is. I t also made it clear my seemingly innate incompatibility with cannabis.